requirements.txt
awsebcli==3.14.2
backcall==0.1.0
beautifulsoup4==4.6.0
boto3==1.7.52
botocore==1.10.52
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
Django==2.1a1
django-cors-headers==2.3.0
django-debug-toolbar==1.9.1
django-storages==1.6.6
django-webpack-loader==0.6.0
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
jmespath==0.9.3
mysql-connector-python==8.0.11
mysqlclient==1.3.12
numpy==1.14.5
openpyxl==2.5.4
parso==0.2.1
pathspec==0.5.5
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.6
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.19.0
s3transfer==0.1.13
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
